security group related methods moved to provider configurator
