+++ provider/openstack/provider.go
@@ -9,7 +9,6 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net/http"
-	"regexp"
 	"strings"
 	"sync"
 	"time"
@@ -242,7 +241,9 @@ hpcloud:
 func (p environProvider) Open(cfg *config.Config) (environs.Environ, error) {
 	logger.Infof("opening environment %q", cfg.Name())
 	e := new(Environ)
-	e.configurator = new(defaultProviderConfigurator)
+	e.configurator = &defaultProviderConfigurator{
+		environ: e,
+	}
 	err := e.SetConfig(cfg)
 	if err != nil {
 		return nil, err
@@ -1063,15 +1064,13 @@ func (e *Environ) StartInstance(args environs.StartInstanceParams) (*environs.St
 
 	cfg := e.Config()
 	var groupNames = make([]nova.SecurityGroupName, 0)
-	if e.configurator.UseSecurityGroups() {
-		groups, err := e.setUpGroups(args.InstanceConfig.MachineId, cfg.APIPort())
-		if err != nil {
-			return nil, fmt.Errorf("cannot set up groups: %v", err)
-		}
+	groups, err := e.configurator.SetUpGroups(args.InstanceConfig.MachineId, cfg.APIPort())
+	if err != nil {
+		return nil, fmt.Errorf("cannot set up groups: %v", err)
+	}
 
-		for _, g := range groups {
-			groupNames = append(groupNames, nova.SecurityGroupName{g.Name})
-		}
+	for _, g := range groups {
+		groupNames = append(groupNames, nova.SecurityGroupName{g.Name})
 	}
 	machineName := resourceName(
 		names.NewMachineTag(args.InstanceConfig.MachineId),
@@ -1141,27 +1140,12 @@ func isNoValidHostsError(err error) bool {
 
 func (e *Environ) StopInstances(ids ...instance.Id) error {
 	// If in instance firewall mode, gather the security group names.
-	var securityGroupNames []string
-	if e.configurator.UseSecurityGroups() {
-		if e.Config().FirewallMode() == config.FwInstance {
-			instances, err := e.Instances(ids)
-			if err == environs.ErrNoInstances {
-				return nil
-			}
-			securityGroupNames = make([]string, 0, len(ids))
-			for _, inst := range instances {
-				if inst == nil {
-					continue
-				}
-				openstackName := inst.(*openstackInstance).getServerDetail().Name
-				lastDashPos := strings.LastIndex(openstackName, "-")
-				if lastDashPos == -1 {
-					return fmt.Errorf("cannot identify machine ID in openstack server name %q", openstackName)
-				}
-				securityGroupName := e.machineGroupName(openstackName[lastDashPos+1:])
-				securityGroupNames = append(securityGroupNames, securityGroupName)
-			}
-		}
+	securityGroupNames, err := e.configurator.GetSecurityGroups(ids...)
+	if err == environs.ErrNoInstances {
+		return nil
+	}
+	if err != nil {
+		return err
 	}
 	logger.Debugf("terminating instances %v", ids)
 	if err := e.terminateInstances(ids); err != nil {
@@ -1322,24 +1306,7 @@ func (e *Environ) Destroy() error {
 	if err != nil {
 		return errors.Trace(err)
 	}
-	novaClient := e.nova()
-	if e.configurator.UseSecurityGroups() {
-		securityGroups, err := novaClient.ListSecurityGroups()
-		if err != nil {
-			return errors.Annotate(err, "cannot list security groups")
-		}
-		re, err := regexp.Compile(fmt.Sprintf("^%s(-\\d+)?$", e.jujuGroupName()))
-		if err != nil {
-			return errors.Trace(err)
-		}
-		globalGroupName := e.globalGroupName()
-		for _, group := range securityGroups {
-			if re.MatchString(group.Name) || group.Name == globalGroupName {
-				deleteSecurityGroup(novaClient, group.Name, group.Id)
-			}
-		}
-	}
-	return nil
+	return e.configurator.DeleteGlobalGroups()
 }
 
 func (e *Environ) globalGroupName() string {
@@ -1451,161 +1418,19 @@ func (e *Environ) portsInGroup(name string) (portRanges []network.PortRange, err
 // TODO: following 30 lines nearly verbatim from environs/ec2
 
 func (e *Environ) OpenPorts(ports []network.PortRange) error {
-	if e.Config().FirewallMode() != config.FwGlobal {
-		return fmt.Errorf("invalid firewall mode %q for opening ports on environment",
-			e.Config().FirewallMode())
-	}
-	if err := e.openPortsInGroup(e.globalGroupName(), ports); err != nil {
-		return err
-	}
-	logger.Infof("opened ports in global group: %v", ports)
-	return nil
+	return e.configurator.OpenPorts(ports)
 }
 
 func (e *Environ) ClosePorts(ports []network.PortRange) error {
-	if e.Config().FirewallMode() != config.FwGlobal {
-		return fmt.Errorf("invalid firewall mode %q for closing ports on environment",
-			e.Config().FirewallMode())
-	}
-	if err := e.closePortsInGroup(e.globalGroupName(), ports); err != nil {
-		return err
-	}
-	logger.Infof("closed ports in global group: %v", ports)
-	return nil
+	return e.configurator.ClosePorts(ports)
 }
 
 func (e *Environ) Ports() ([]network.PortRange, error) {
-	if e.Config().FirewallMode() != config.FwGlobal {
-		return nil, fmt.Errorf("invalid firewall mode %q for retrieving ports from environment",
-			e.Config().FirewallMode())
-	}
-	return e.portsInGroup(e.globalGroupName())
+	return e.configurator.Ports()
 }
 
 func (e *Environ) Provider() environs.EnvironProvider {
 	return &providerInstance
-}
-
-func (e *Environ) setUpGlobalGroup(groupName string, apiPort int) (nova.SecurityGroup, error) {
-	return e.ensureGroup(groupName,
-		[]nova.RuleInfo{
-			{
-				IPProtocol: "tcp",
-				FromPort:   22,
-				ToPort:     22,
-				Cidr:       "0.0.0.0/0",
-			},
-			{
-				IPProtocol: "tcp",
-				FromPort:   apiPort,
-				ToPort:     apiPort,
-				Cidr:       "0.0.0.0/0",
-			},
-			{
-				IPProtocol: "tcp",
-				FromPort:   1,
-				ToPort:     65535,
-			},
-			{
-				IPProtocol: "udp",
-				FromPort:   1,
-				ToPort:     65535,
-			},
-			{
-				IPProtocol: "icmp",
-				FromPort:   -1,
-				ToPort:     -1,
-			},
-		})
-}
-
-// setUpGroups creates the security groups for the new machine, and
-// returns them.
-//
-// Instances are tagged with a group so they can be distinguished from
-// other instances that might be running on the same OpenStack account.
-// In addition, a specific machine security group is created for each
-// machine, so that its firewall rules can be configured per machine.
-//
-// Note: ideally we'd have a better way to determine group membership so that 2
-// people that happen to share an openstack account and name their environment
-// "openstack" don't end up destroying each other's machines.
-func (e *Environ) setUpGroups(machineId string, apiPort int) ([]nova.SecurityGroup, error) {
-	jujuGroup, err := e.setUpGlobalGroup(e.jujuGroupName(), apiPort)
-	if err != nil {
-		return nil, err
-	}
-	var machineGroup nova.SecurityGroup
-	switch e.Config().FirewallMode() {
-	case config.FwInstance:
-		machineGroup, err = e.ensureGroup(e.machineGroupName(machineId), nil)
-	case config.FwGlobal:
-		machineGroup, err = e.ensureGroup(e.globalGroupName(), nil)
-	}
-	if err != nil {
-		return nil, err
-	}
-	groups := []nova.SecurityGroup{jujuGroup, machineGroup}
-	if e.ecfg().useDefaultSecurityGroup() {
-		defaultGroup, err := e.nova().SecurityGroupByName("default")
-		if err != nil {
-			return nil, fmt.Errorf("loading default security group: %v", err)
-		}
-		groups = append(groups, *defaultGroup)
-	}
-	return groups, nil
-}
-
-// zeroGroup holds the zero security group.
-var zeroGroup nova.SecurityGroup
-
-// ensureGroup returns the security group with name and perms.
-// If a group with name does not exist, one will be created.
-// If it exists, its permissions are set to perms.
-func (e *Environ) ensureGroup(name string, rules []nova.RuleInfo) (nova.SecurityGroup, error) {
-	novaClient := e.nova()
-	// First attempt to look up an existing group by name.
-	group, err := novaClient.SecurityGroupByName(name)
-	if err == nil {
-		// Group exists, so assume it is correctly set up and return it.
-		// TODO(jam): 2013-09-18 http://pad.lv/121795
-		// We really should verify the group is set up correctly,
-		// because deleting and re-creating environments can get us bad
-		// groups (especially if they were set up under Python)
-		return *group, nil
-	}
-	// Doesn't exist, so try and create it.
-	group, err = novaClient.CreateSecurityGroup(name, "juju group")
-	if err != nil {
-		if !gooseerrors.IsDuplicateValue(err) {
-			return zeroGroup, err
-		} else {
-			// We just tried to create a duplicate group, so load the existing group.
-			group, err = novaClient.SecurityGroupByName(name)
-			if err != nil {
-				return zeroGroup, err
-			}
-			return *group, nil
-		}
-	}
-	// The new group is created so now add the rules.
-	group.Rules = make([]nova.SecurityGroupRule, len(rules))
-	for i, rule := range rules {
-		rule.ParentGroupId = group.Id
-		if rule.Cidr == "" {
-			// http://pad.lv/1226996 Rules that don't have a CIDR
-			// are meant to apply only to this group. If you don't
-			// supply CIDR or GroupId then openstack assumes you
-			// mean CIDR=0.0.0.0/0
-			rule.GroupId = &group.Id
-		}
-		groupRule, err := novaClient.CreateSecurityGroupRule(rule)
-		if err != nil && !gooseerrors.IsDuplicateValue(err) {
-			return zeroGroup, err
-		}
-		group.Rules[i] = *groupRule
-	}
-	return *group, nil
 }
 
 // deleteSecurityGroups deletes the given security groups. If a security
@@ -1626,39 +1451,6 @@ func (e *Environ) deleteSecurityGroups(securityGroupNames []string) error {
 		}
 	}
 	return nil
-}
-
-// deleteSecurityGroup attempts to delete the security group. Should it fail,
-// the deletion is retried due to timing issues in openstack. A security group
-// cannot be deleted while it is in use. Theoretically we terminate all the
-// instances before we attempt to delete the associated security groups, but
-// in practice nova hasn't always finished with the instance before it
-// returns, so there is a race condition where we think the instance is
-// terminated and hence attempt to delete the security groups but nova still
-// has it around internally. To attempt to catch this timing issue, deletion
-// of the groups is tried multiple times.
-func deleteSecurityGroup(novaclient *nova.Client, name, id string) {
-	attempts := utils.AttemptStrategy{
-		Total: 30 * time.Second,
-		Delay: time.Second,
-	}
-	logger.Debugf("deleting security group %q", name)
-	i := 0
-	for attempt := attempts.Start(); attempt.Next(); {
-		err := novaclient.DeleteSecurityGroup(id)
-		if err == nil {
-			return
-		}
-		i++
-		if i%4 == 0 {
-			message := fmt.Sprintf("waiting to delete security group %q", name)
-			if i != 4 {
-				message = "still " + message
-			}
-			logger.Debugf(message)
-		}
-	}
-	logger.Warningf("cannot delete security group %q. Used by another environment?", name)
 }
 
 func (e *Environ) terminateInstances(ids []instance.Id) error {
